+++ services/auth-api/app/models/balance.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, ForeignKey, Numeric
+from sqlalchemy import Column, BigInteger, ForeignKey, Numeric
 
 from app.db.base import Base
 
@@ -6,5 +6,5 @@ from app.db.base import Base
 class Balance(Base):
     __tablename__ = "balances"
 
-    user_id = Column(Integer, ForeignKey("users.id"), primary_key=True)
+    user_id = Column(BigInteger, ForeignKey("users.id"), primary_key=True)
     amount = Column(Numeric(12,2), nullable = False, default = 0)
